pyproject.toml
[project]
name = "my-gadk-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.18.0",
    "litellm>=1.79.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "timezonefinder>=8.1.0",
]
